requirements.txt
python-telegram-bot==22.7
python-dotenv==1.0.1
requests==2.32.3
